Process_Image/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
Pillow==10.4.0
numpy==1.26.4
piexif==1.1.3
face-recognition==1.3.0
geopy==2.4.1
transformers==4.49.0
spacy==3.8.3
torch